+++ centml/compiler/prediction/profiler.py
@@ -2,6 +2,7 @@ import ast
 import csv
 import logging
 from typing import Dict
+
 import torch
 import torch.fx
 from sklearn.neighbors import KDTree  # type: ignore
